// src/app/api/model/productionCompany.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductionCompany { 
    id?: number;
    logo_path?: string | null;
    name?: string;
    origin_country?: string;
}